refactor : DTO, VO 유효성 검증 수정, 데이터 타입 변경₩
1. request DTO : NotNull, 최대, 최소 값 검증 추가, 불필요 필드 제거
2. response DTO : todoId long 타입으로 변경
3. VO : Reference 타입, Long 으로 변경

+++ src/main/java/arraysort/todolist/domain/TodoListDto.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @Builder
 public class TodoListDto {
 
-    private int todoId;
+    private long todoId;
 
     private String userId;
